Fix kerning bug for nina tile on Safari mobile
Seems to be a reasonably well known bug within mobile safari that renders faux styles incorrectly; http://stackoverflow.com/a/11879044. This occurs when a font-style of bold is used on a TTF that does not have a bold style. Instead of defining a weight of bold, use the medium font that is present on the page.